+++ activemodel/test/cases/mass_assignment_test.rb
@@ -1,26 +1,24 @@
 require 'cases/helper'
 
-class ModelTest < ActiveModel::TestCase
-  include ActiveModel::Lint::Tests
-
-  class BasicModel
-    include ActiveModel::Model
+class MassAssignmentTest < ActiveModel::TestCase
+  class Model
+    include ActiveModel::MassAssignment
     attr_accessor :attr
   end
 
   def setup
-    @model = BasicModel.new
+    @model = Model.new
   end
 
   def test_initialize_with_params
-    object = BasicModel.new(:attr => "value")
+    object = Model.new(:attr => "value")
     assert_equal object.attr, "value"
   end
 
   def test_initialize_with_nil_or_empty_hash_params_does_not_explode
     assert_nothing_raised do
-      BasicModel.new()
-      BasicModel.new({})
+      Model.new()
+      Model.new({})
     end
   end
 end